Remove the perf tests duplicated
GearVRf-DCO-1.0-Signed-off-by: Eliamara Silva <eliamara.s@samsung.com>
